# [FIX] stock_sale: improve performance of sale_order_ids computation