Improve _compute_sale_order_ids performances by saving
stock.move.line search results to defaultdict before going
through records in self.


##### speedup

DB: 1939 lots, 110k stock_moves, 100k move_lines
Reported Average Time of
```python 
lots.modified(fnames=['name'])
lots.recompute()
```

| len(lots) | before PR | after PR |
|:--------------------:|:-------------:|:------------:|
| 50 | 1.61s | 0.05s |
| 200 | 5s | 0.09s |
| 500 | 11.29s | 0.18s |
| 1000 | 21.46s | 0.36s |
| 1939 | 40.7s | 0.69s |

This is a backport of https://github.com/odoo/odoo/pull/71949 which could have been made against V13.
I slightly altered the code so that it never fails for the `stock.group_stock_user` user due to right issues either.